requirements.txt
pip==24.3.1
openai==1.57.0
pillow==11.0.0
aiohttp==3.11.10
aiofiles==24.1.0
loguru==0.7.3